pyproject.toml
[tool.poetry]
name = "jocko-willink-bot"
version = "0.1.0"
description = ""
authors = ["Wookyung Jeong <77jwk0724@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "^2.4.0"
pillow = "^10.4.0"
piexif = "^1.1.3"

[tool.poetry.group.dev.dependencies]
pyright = "^1.1.372"
ruff = "^0.5.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
